// src/app/shared/services/swagger/model/security.ts
/**
 * Demo API
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { Central } from './central';

export interface Security {
    securityId?: number;
    userId?: number;
    windowsIdentityUserName?: string;
    windowsIdentityGroupName?: string;
    roleId?: number;
    centralId?: number;
    securityUserGroupId?: number;
    central?: Central;
}